Delaunay triangulation works! There's a bug with merging lines sometimes, which I'll ignore

+++ delaunay_line.cpp
@@ -45,10 +45,10 @@ bool DelaunayLine::doesCrossLine(const DelaunayLine &otherLine)
          (unitCD.cross(unitDA)[2] * unitCD.cross(unitDB)[2] < 0) )
         return true;
 
-    Eigen::Vector3i A{m_x1, m_y1, 0};
-    Eigen::Vector3i B{m_x2, m_y2, 0};
-    Eigen::Vector3i C{otherLine.m_x1, otherLine.m_y1, 0};
-    Eigen::Vector3i D{otherLine.m_x2, otherLine.m_y2, 0};
+    Eigen::Vector3d A{m_x1, m_y1, 0};
+    Eigen::Vector3d B{m_x2, m_y2, 0};
+    Eigen::Vector3d C{otherLine.m_x1, otherLine.m_y1, 0};
+    Eigen::Vector3d D{otherLine.m_x2, otherLine.m_y2, 0};
 
     if ( onSegment(C, A, B) || onSegment(D, A, B) || onSegment(A, C, D) || onSegment(B, C, D) )
         return true;
@@ -101,13 +101,13 @@ bool valueIsBetween(int testValue, int side1, int side2)
 }
 
 
-bool onSegment(Eigen::Vector3i testPoint, Eigen::Vector3i side1, Eigen::Vector3i side2)
+bool onSegment(Eigen::Vector3d testPoint, Eigen::Vector3d side1, Eigen::Vector3d side2)
 {
     double lengthThroughTestPoint = (testPoint - side1).norm() +
                                     (side2 - testPoint).norm();
     double segmentLength = (side2 - side1).norm();
 
-    // Does count sharing an endpoint as being "on the segment"
+    // Doesn't count sharing an endpoint as being "on the segment"
     if ( almostEqual( (testPoint - side1).norm(), 0.0 ) ||
          almostEqual( (side2 - testPoint).norm(), 0.0 ) )
     {
